Tried to compile the static helper lib but it still does not work
