--- YUViewUnitTest/helper2/helper.pro
@@ -1,11 +1,0 @@
-QT += core gui widgets opengl xml concurrent network
-
-TEMPLATE = lib
-CONFIG += staticlib
-CONFIG += c++1z
-CONFIG -= debug_and_release
-CONFIG += object_parallel_to_source
-
-SOURCES += $$files(*.cpp, true)
-HEADERS += $$files(*.h, true)
-FORMS += $$files(*.ui, false)
