+++ YUViewUnitTest/video/pixelFormat/pixelFormat.pro
@@ -1,0 +1,8 @@
+TEMPLATE = subdirs
+
+requires(qtHaveModule(testlib))
+
+SUBDIRS = pixelFormatYUV \
+          pixelFormatRGB \
+          pixelFormatYUVGuess \
+          pixelFormatRGBGuess
