+++ YUViewUnitTest/video/videoHandler/videoHandler.pro
@@ -1,0 +1,6 @@
+TEMPLATE = subdirs
+
+requires(qtHaveModule(testlib))
+
+SUBDIRS = videoHandlerDifference \
+          videoHandlerYUV
